feat(demo): remove demo mode
moved to a separate plugin

+++ inc/geolocation.class.php
@@ -247,8 +247,10 @@ class PluginStorkmdmGeolocation extends CommonDBTM {
       return $tab;
    }
 
-   public static function hook_computer_purge(CommonDBTM $item) {
-      $geolocation = new self();
-      $geolocation->deleteByCriteria(array('computers_id' => $item->getID()));
+   public function hook_computer_purge(CommonDBTM $item) {
+      if ($item instanceof Computer) {
+         $geolocation = new self();
+         $geolocation->deleteByCriteria(array('computers_id' => $item->getID()));
+      }
    }
 }
